# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberNexus AI - Cybersecurity Telemetry & Fraud Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /telemetry/events:
    get:
      operationId: getTelemetryEvents
      tags: [telemetry]
      summary: List cybersecurity telemetry events
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of telemetry events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelemetryEvent"
  /telemetry/events/{id}:
    get:
      operationId: getTelemetryEvent
      tags: [telemetry]
      summary: Get a specific telemetry event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Telemetry event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryEvent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: List correlated threat alerts (role-filtered)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, true_positive, false_positive, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
...
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get a specific alert with full XAI explanation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert with explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /alerts/{id}/feedback:
    patch:
      operationId: submitAlertFeedback
      tags: [alerts]
      summary: Analyst submits TP/FP feedback for an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertFeedbackInput"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "403":
          description: Forbidden - analysts only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /correlations:
    get:
      operationId: getCorrelations
      tags: [correlations]
      summary: List AI correlation events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of correlations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Correlation"
  /correlations/{id}:
    get:
      operationId: getCorrelation
      tags: [correlations]
      summary: Get a specific correlation with XAI explanation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Correlation with explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrelationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, analyst, user]
        department:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, name, role, createdAt]
    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          type: string
          enum: [analyst, user]
        department:
          type: string
      required: [email, password, name]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    TelemetryEvent:
      type: object
      properties:
        id:
          type: integer
        eventType:
          type: string
          enum: [login_attempt, network_anomaly, endpoint_alert, certificate_anomaly, privilege_escalation, lateral_movement, data_exfiltration]
        severity:
          type: string
          enum: [critical, high, medium, low]
        sourceIp:
          type: string
        destinationIp:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        deviceId:
          type: ["string", "null"]
        description:
          type: string
        timestamp:
          type: string
        isCorrelated:
          type: boolean
      required: [id, eventType, severity, sourceIp, description, timestamp, isCorrelated]
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountId:
          type: string
        transactionType:
          type: string
          enum: [transfer, withdrawal, deposit, payment, wire]
        amount:
          type: number
        currency:
          type: string
        merchantName:
          type: ["string", "null"]
        merchantCategory:
          type: ["string", "null"]
        riskScore:
          type: number
        isAnomaly:
          type: boolean
        anomalyReasons:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [approved, flagged, blocked, under_review]
        timestamp:
          type: string
      required: [id, userId, accountId, transactionType, amount, currency, riskScore, isAnomaly, anomalyReasons, status, timestamp]
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        category:
          type: string
          enum: [fraud, cyber_threat, quantum_risk, insider_threat, data_breach, account_takeover]
        status:
          type: string
          enum: [open, true_positive, false_positive, resolved]
        riskScore:
          type: number
        affectedUserId:
          type: ["integer", "null"]
        affectedAccountId:
          type: ["string", "null"]
        analystId:
          type: ["integer", "null"]
        analystNote:
          type: ["string", "null"]
        telemetryEventIds:
          type: array
          items:
            type: integer
        transactionIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, description, severity, category, status, riskScore, telemetryEventIds, transactionIds, createdAt, updatedAt]
    AlertDetail:
      type: object
      properties:
        alert:
          $ref: "#/components/schemas/Alert"
        xaiExplanation:
          $ref: "#/components/schemas/XAIExplanation"
        relatedTelemetry:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryEvent"
        relatedTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        correlationPath:
          type: array
          items:
            type: string
      required: [alert, xaiExplanation, relatedTelemetry, relatedTransactions, correlationPath]
    AlertFeedbackInput:
      type: object
      properties:
        status:
          type: string
          enum: [true_positive, false_positive, resolved]
        note:
          type: string
      required: [status]
...
    XAIExplanation:
      type: object
      properties:
        summary:
          type: string
        confidence:
          type: number
        riskFactors:
          type: array
          items:
            $ref: "#/components/schemas/RiskFactor"
        modelUsed:
          type: string
        graphFeatures:
          type: array
          items:
            type: string
        recommendedAction:
          type: string
      required: [summary, confidence, riskFactors, modelUsed, graphFeatures, recommendedAction]
    RiskFactor:
      type: object
      properties:
        factor:
          type: string
        weight:
          type: number
        description:
          type: string
      required: [factor, weight, description]
    Correlation:
      type: object
      properties:
        id:
          type: integer
        correlationType:
          type: string
          enum: [telemetry_transaction, multi_account, device_cluster, geo_anomaly, time_pattern, quantum_exposure]
        strength:
          type: number
        involvedEntities:
          type: array
          items:
            type: string
        telemetryEventIds:
          type: array
          items:
            type: integer
        transactionIds:
          type: array
          items:
            type: integer
        alertId:
          type: ["integer", "null"]
        description:
          type: string
        detectedAt:
          type: string
      required: [id, correlationType, strength, involvedEntities, telemetryEventIds, transactionIds, description, detectedAt]
    CorrelationDetail:
      type: object
      properties:
        correlation:
          $ref: "#/components/schemas/Correlation"
        xaiExplanation:
          $ref: "#/components/schemas/XAIExplanation"
        graphData:
          $ref: "#/components/schemas/GraphData"
      required: [correlation, xaiExplanation, graphData]
    GraphData:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
      required: [nodes, edges]
    GraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [user, device, account, transaction, ip_address, alert]
        riskScore:
          type: number
        x:
          type: number
        y:
          type: number
        z:
          type: number
      required: [id, label, type, riskScore, x, y, z]
    GraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        weight:
          type: number
        edgeType:
          type: string
          enum: [transacted_with, accessed_from, correlated, flagged, owns]
      required: [source, target, weight, edgeType]
...