# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CyberNexus AI - Cybersecurity Telemetry & Fraud Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  error: string;
}