# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CyberNexus AI - Cybersecurity Telemetry & Fraud Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  department?: string | null;
  createdAt: string;
}