# lib/api-zod/src/generated/types/xAIExplanation.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CyberNexus AI - Cybersecurity Telemetry & Fraud Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFactor } from './riskFactor';

export interface XAIExplanation {
  summary: string;
  confidence: number;
  riskFactors: RiskFactor[];
  modelUsed: string;
  graphFeatures: string[];
  recommendedAction: string;
}